requirements.txt
torch
torch-geometric
transformers
scikit-learn
Pillow
networkx
matplotlib
torchvision